Speed up doc build by making read and write safe
