+++ build_tools/circle/doc_environment.yml
@@ -30,7 +30,6 @@ dependencies:
   - sphinx-gallery
   - sphinx-copybutton
   - numpydoc
-  - sphinx-prompt
   - plotly
   - polars
   - pooch
