+++ build_tools/update_environments_and_lock_files.py
@@ -349,7 +349,6 @@ build_metadata_list = [
             "sphinx-gallery",
             "sphinx-copybutton",
             "numpydoc",
-            "sphinx-prompt",
             "plotly",
             "polars",
             "pooch",
@@ -376,7 +375,6 @@ build_metadata_list = [
             "sphinx-gallery": "min",
             "sphinx-copybutton": "min",
             "numpydoc": "min",
-            "sphinx-prompt": "min",
             "sphinxext-opengraph": "min",
             "plotly": "min",
             "polars": "min",
@@ -406,7 +404,6 @@ build_metadata_list = [
             "sphinx-gallery",
             "sphinx-copybutton",
             "numpydoc",
-            "sphinx-prompt",
             "plotly",
             "polars",
             "pooch",
