+++ doc/developers/misc_info.rst
@@ -25,31 +25,31 @@ as `/usr/bin/gcc`, it does not directly support OpenMP.
 
 To check if `scikit-learn` has been built correctly with OpenMP, run
 
-.. prompt:: bash $
+.. code-block:: console
 
-  python -c "import sklearn; sklearn.show_versions()"
+  $ python -c "import sklearn; sklearn.show_versions()"
 
 and check if it contains `Built with OpenMP: True`.
 
 When using conda on Mac, you can also check that the custom compilers
 are properly installed from conda-forge using the following command:
 
-.. prompt:: bash $
+.. code-block:: console
 
-    conda list
+    $ conda list
 
 which should include ``compilers`` and ``llvm-openmp``.
 
 The compilers meta-package will automatically set custom environment
 variables:
 
-.. prompt:: bash $
+.. code-block:: console
 
-    echo $CC
-    echo $CXX
-    echo $CFLAGS
-    echo $CXXFLAGS
-    echo $LDFLAGS
+    $ echo $CC
+    $ echo $CXX
+    $ echo $CFLAGS
+    $ echo $CXXFLAGS
+    $ echo $LDFLAGS
 
 They point to files and folders from your ``sklearn-dev`` conda environment
 (in particular in the `bin/`, `include/` and `lib/` subfolders). For instance
@@ -72,9 +72,9 @@ Note on dependencies for other Linux distributions
 When precompiled wheels of the runtime dependencies are not available for your
 architecture (e.g. **ARM**), you can install the system versions:
 
-.. prompt::
+.. code-block:: console
 
-  sudo apt-get install cython3 python3-numpy python3-scipy
+  $ sudo apt-get install cython3 python3-numpy python3-scipy
 
 
 Notes on Meson
@@ -85,6 +85,6 @@ scikit-learn installations and meson builds can lead to conflicts.
 You can use the `Makefile` provided in the `scikit-learn repository <https://github.com/scikit-learn/scikit-learn/>`__
 to remove conflicting builds by calling:
 
-.. prompt:: bash $
+.. code-block:: console
 
-    make clean
+    $ make clean
