+++ doc/install.rst
@@ -73,19 +73,19 @@ Installing the latest release
           Note that the virtual environment is optional but strongly recommended, in
           order to avoid potential conflicts with other packages.
 
-          .. prompt:: powershell
+          .. code-block:: ps1con
 
-            python -m venv sklearn-env
-            sklearn-env\Scripts\activate  # activate
-            pip install -U scikit-learn
+            PS C:\> python -m venv sklearn-env
+            PS C:\> sklearn-env\Scripts\activate  # activate
+            PS C:\> pip install -U scikit-learn
 
           In order to check your installation, you can use:
 
-          .. prompt:: powershell
+          .. code-block:: ps1con
 
-            python -m pip show scikit-learn  # show scikit-learn version and location
-            python -m pip freeze             # show all installed packages in the environment
-            python -c "import sklearn; sklearn.show_versions()"
+            PS C:\> python -m pip show scikit-learn  # show scikit-learn version and location
+            PS C:\> python -m pip freeze             # show all installed packages in the environment
+            PS C:\> python -c "import sklearn; sklearn.show_versions()"
 
         .. tab-item:: conda
           :class-label: tab-6
@@ -112,19 +112,19 @@ Installing the latest release
           Note that the virtual environment is optional but strongly recommended, in
           order to avoid potential conflicts with other packages.
 
-          .. prompt:: bash
+          .. code-block:: console
 
-            python -m venv sklearn-env
-            source sklearn-env/bin/activate  # activate
-            pip install -U scikit-learn
+            $ python -m venv sklearn-env
+            $ source sklearn-env/bin/activate  # activate
+            $ pip install -U scikit-learn
 
           In order to check your installation, you can use:
 
-          .. prompt:: bash
+          .. code-block:: console
 
-            python -m pip show scikit-learn  # show scikit-learn version and location
-            python -m pip freeze             # show all installed packages in the environment
-            python -c "import sklearn; sklearn.show_versions()"
+            $ python -m pip show scikit-learn  # show scikit-learn version and location
+            $ python -m pip freeze             # show all installed packages in the environment
+            $ python -c "import sklearn; sklearn.show_versions()"
 
         .. tab-item:: conda
           :class-label: tab-6
@@ -145,10 +145,10 @@ Installing the latest release
           Python 3 is usually installed by default on most Linux distributions. To
           check if you have it installed, try:
 
-          .. prompt:: bash
+          .. code-block:: console
 
-            python3 --version
-            pip3 --version
+            $ python3 --version
+            $ pip3 --version
 
           If you don't have Python 3 installed, please install `python3` and
           `python3-pip` from your distribution's package manager.
@@ -158,19 +158,19 @@ Installing the latest release
           Note that the virtual environment is optional but strongly recommended, in
           order to avoid potential conflicts with other packages.
 
-          .. prompt:: bash
+          .. code-block:: console
 
-            python3 -m venv sklearn-env
-            source sklearn-env/bin/activate  # activate
-            pip3 install -U scikit-learn
+            $ python3 -m venv sklearn-env
+            $ source sklearn-env/bin/activate  # activate
+            $ pip3 install -U scikit-learn
 
           In order to check your installation, you can use:
 
-          .. prompt:: bash
+          .. code-block:: console
 
-            python3 -m pip show scikit-learn  # show scikit-learn version and location
-            python3 -m pip freeze             # show all installed packages in the environment
-            python3 -c "import sklearn; sklearn.show_versions()"
+            $ python3 -m pip show scikit-learn  # show scikit-learn version and location
+            $ python3 -m pip freeze             # show all installed packages in the environment
+            $ python3 -c "import sklearn; sklearn.show_versions()"
 
         .. tab-item:: conda
           :class-label: tab-6
@@ -247,9 +247,9 @@ Alpine Linux's package is provided through the `official repositories
 ``py3-scikit-learn`` for Python.
 It can be installed by typing the following command:
 
-.. prompt:: bash
+.. code-block:: console
 
-  sudo apk add py3-scikit-learn
+  $ sudo apk add py3-scikit-learn
 
 
 Arch Linux
@@ -260,9 +260,9 @@ Arch Linux's package is provided through the `official repositories
 ``python-scikit-learn`` for Python.
 It can be installed by typing the following command:
 
-.. prompt:: bash
+.. code-block:: console
 
-  sudo pacman -S python-scikit-learn
+  $ sudo pacman -S python-scikit-learn
 
 
 Debian/Ubuntu
@@ -275,9 +275,9 @@ Note that scikit-learn requires Python 3, hence the need to use the `python3-`
 suffixed package names.
 Packages can be installed using ``apt-get``:
 
-.. prompt:: bash
+.. code-block:: console
 
-  sudo apt-get install python3-sklearn python3-sklearn-lib python-sklearn-doc
+  $ sudo apt-get install python3-sklearn python3-sklearn-lib python-sklearn-doc
 
 
 Fedora
@@ -287,9 +287,9 @@ The Fedora package is called ``python3-scikit-learn`` for the python 3 version,
 the only one available in Fedora.
 It can be installed using ``dnf``:
 
-.. prompt:: bash
+.. code-block:: console
 
-  sudo dnf install python3-scikit-learn
+  $ sudo dnf install python3-scikit-learn
 
 
 NetBSD
@@ -307,9 +307,9 @@ where ``XY`` denotes the Python version.
 It can be installed by typing the following
 command:
 
-.. prompt:: bash
+.. code-block:: console
 
-  sudo port install py312-scikit-learn
+  $ sudo port install py312-scikit-learn
 
 
 Anaconda and Enthought Deployment Manager for all supported platforms
@@ -329,9 +329,9 @@ Intel Extension for Scikit-learn
 Intel maintains an optimized x86_64 package, available in PyPI (via `pip`),
 and in the `main`, `conda-forge` and `intel` conda channels:
 
-.. prompt:: bash
+.. code-block:: console
 
-  conda install scikit-learn-intelex
+  $ conda install scikit-learn-intelex
 
 This package has an Intel optimized version of many estimators. Whenever
 an alternative implementation doesn't exist, scikit-learn implementation
@@ -343,9 +343,9 @@ Note that those solvers are not enabled by default, please refer to the
 `scikit-learn-intelex <https://uxlfoundation.github.io/scikit-learn-intelex/latest/patching.html>`_
 documentation for more details on usage scenarios. Direct export example:
 
-.. prompt:: python >>>
+.. code-block:: pycon
 
-  from sklearnex.neighbors import NearestNeighbors
+  >>> from sklearnex.neighbors import NearestNeighbors
 
 Compatibility with the standard scikit-learn solvers is checked by running the
 full scikit-learn test suite via automated continuous integration as reported
@@ -397,9 +397,9 @@ using the ``regedit`` tool:
 
 #. Reinstall scikit-learn (ignoring the previous broken installation):
 
-   .. prompt:: powershell
+   .. code-block:: ps1con
 
-      pip install --exists-action=i scikit-learn
+      PS C:\> pip install --exists-action=i scikit-learn
 
 
 .. _install_nightly_builds:
@@ -414,9 +414,9 @@ basis.
 You can install the nightly build of scikit-learn using the `scientific-python-nightly-wheels`
 index from the PyPI registry of `anaconda.org`:
 
-.. prompt:: bash $
+.. code-block:: console
 
-  pip install --pre --extra-index https://pypi.anaconda.org/scientific-python-nightly-wheels/simple scikit-learn
+  $ pip install --pre --extra-index https://pypi.anaconda.org/scientific-python-nightly-wheels/simple scikit-learn
 
 Note that first uninstalling scikit-learn might be required to be able to
 install nightly builds of scikit-learn.
