+++ doc/sphinxext/dropdown_anchors.py
@@ -56,3 +56,8 @@ class DropdownAnchorAdder(SphinxPostTransform):
 
 def setup(app):
     app.add_post_transform(DropdownAnchorAdder)
+    return {
+        "version": "0.1",
+        "parallel_read_safe": True,
+        "parallel_write_safe": True,
+    }
